implement $now() and $currentDate() dialects for getting current date
